requirements.txt
Django==2.2.6
django-emoji-picker==0.0.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Pillow==6.2.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
